% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{PCAVisualization}
\alias{PCAVisualization}
\title{PCAVisualization}
\usage{
PCAVisualization(predictSet, predictResult)
}
\arguments{
\item{predictSet}{Data for stemscore prediction.}

\item{predictResult}{The result from scorePrediction function}
}
\value{
a plot of PCA dimentional reduction
}
\description{
Visualize the prediction result with PCA to check if the predicted stem cell can be distinguished from non-stem cell.
}
