% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{readMatrix}
\alias{readMatrix}
\title{readMatrix}
\usage{
readMatrix(dir)
}
\arguments{
\item{dir}{the path of the matrix to be read}
}
\value{
list object of the read matrix
}
\description{
Read data for training
}
