% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{scorePrediction}
\alias{scorePrediction}
\title{scorePrediction}
\usage{
scorePrediction(algorithm, trainSet, predictSet, target)
}
\arguments{
\item{algorithm}{the algorithm used for learning. 'SVM','randomForest','naiveBayes' are supported.}

\item{trainSet}{the data for training. With stem state labelled.}

\item{predictSet}{the data for predictiton.}

\item{target}{the label column of the trainSet}
}
\value{
prediction result in the form of list
}
\description{
Predict the stemscore of predictSet with trainSet and selected machine learning algorithm.
}
